change table creations + date changes

+++ departmentTable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE department (
-        dID INT(1) Primary Key NOT NULL,
-        dName VARCHAR(25) NOT NULL,
+        dID numeric(1) Primary Key NOT NULL,
+        dName VARCHAR(25) NOT NULL
 );
